add CRUD JDBC and upgrade repository tests, refactor domain, remove equals on createdDateTime Add jdbc utils, extractors and exceptions, remove authenticationController(move to UserController), update context, add ini.sql, persistence.properties and sqlqueries.properies add signin and signup jsp, edit jsp links.

+++ src/main/java/com/vaka/service/impl/ReservationServiceImpl.java
@@ -1,15 +1,18 @@
 package com.vaka.service.impl;
 
+import com.vaka.context.ApplicationContext;
 import com.vaka.domain.*;
 import com.vaka.repository.ReservationRepository;
 import com.vaka.repository.RoomRepository;
 import com.vaka.service.ReservationService;
-import com.vaka.util.ApplicationContext;
 import com.vaka.util.DateUtil;
+import com.vaka.util.exception.AuthorizationException;
 import com.vaka.util.exception.NotFoundException;
 import com.vaka.util.exception.ReservationCreationException;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Created by Iaroslav on 11/27/2016.
@@ -19,60 +22,73 @@ public class ReservationServiceImpl implements ReservationService {
     private RoomRepository roomRepository;
 
     @Override
-    public Reservation create(Reservation entity) {
-            return getReservationRepository().create(entity);
-
+    public Reservation create(User loggedUser, Reservation entity) {
+        entity.setCreatedDatetime(LocalDateTime.now());
+        return getReservationRepository().create(entity);
     }
 
     @Override
-    public List<Reservation> findRequested() {
-        return getReservationRepository().findRequested();
+    public List<Reservation> findByStatus(User loggedUser, ReservationStatus status) {
+        if (loggedUser.getRole() != Role.MANAGER)
+            throw new AuthorizationException("Not Allowed.");
+        return getReservationRepository().findByStatus(status);
     }
 
     @Override
-    public List<Reservation> findConfirmed() {
-        return getReservationRepository().findConfirmed();
-    }
+    public Reservation applyRoomForReservation(User loggedUser, Integer roomId, Integer reservationId) {
+        if (loggedUser.getRole() != Role.MANAGER)
+            throw new AuthorizationException("Not Allowed.");
+        Optional<Room> room = getRoomRepository().getById(roomId);
+        Optional<Reservation> request = getReservationRepository().getById(reservationId);
 
-    @Override
-    public Reservation applyRoomForReservation(Integer roomId, Integer reservationId) {
-        Room room = getRoomRepository().getById(roomId);
-        Reservation request = getReservationRepository().getById(reservationId);
-
-        if (room == null || request == null)
+        if (!room.isPresent() || !request.isPresent())
             throw new NotFoundException(String.format("Not found room or request by given id, founded room: %s, request: %s", room, request));
-        boolean datesOverlap = reservationRepository.findConfirmedByRoomId(roomId).stream().filter(//TODO get boolean from bd, existsByRoomIdAndDates()
+        boolean datesOverlap = reservationRepository.findByRoomIdAndStatus(roomId, ReservationStatus.CONFIRMED).stream().filter(//TODO get boolean from bd, existsByRoomIdAndDates()
                 reservation -> DateUtil.areDatesOverlap(reservation.getArrivalDate(),
-                        reservation.getDepartureDate(), request.getArrivalDate(), request.getDepartureDate())
+                        reservation.getDepartureDate(), request.get().getArrivalDate(), request.get().getDepartureDate())
         ).count() > 0;
         if (datesOverlap)
             throw new ReservationCreationException("Requested Reservation arrival or departure dates with active reservations dates conflict");//TODO look rrom confirmed reservations
         //TODO implement isolation(synchronization)
-        request.setRoom(room);
-        request.setStatus(ReservationStatus.CONFIRMED);
-        return update(reservationId, request);
+        request.get().setRoom(room.get());
+        request.get().setStatus(ReservationStatus.CONFIRMED);
+        update(loggedUser, reservationId, request.get());
+        return request.get();//TODO change to boolean
     }
 
     @Override
-    public Reservation getById(Integer id) {
-            return getReservationRepository().getById(id);
+    public Optional<Reservation> getById(User loggedUser, Integer id) {
+        Optional<Reservation> reservation = getReservationRepository().getById(id);
+        if (loggedUser.getRole() == Role.MANAGER) {
+            return reservation;
+        }
+        if (reservation.isPresent()) {
+            if (reservation.get().getUser().getId().equals(loggedUser.getId())) {
+                return reservation;
+            }
+        }
+        throw new AuthorizationException("Not Allowed.");
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(User loggedUser, Integer id) {
+        if (loggedUser.getRole() == Role.MANAGER) {
             return getReservationRepository().delete(id);
+        } else throw new AuthorizationException("Not Allowed.");
     }
 
     @Override
-    public Reservation update(Integer id, Reservation entity) {
+    public boolean update(User loggedUser, Integer id, Reservation entity) {
+        if (loggedUser.getRole() == Role.MANAGER) {
             return getReservationRepository().update(id, entity);
+        } else throw new AuthorizationException("Not Allowed.");
     }
 
     private ReservationRepository getReservationRepository() {
         if (reservationRepository == null) {
             synchronized (this) {
                 if (reservationRepository == null)
-                    reservationRepository = ApplicationContext.getBean(ReservationRepository.class);
+                    reservationRepository = ApplicationContext.getInstance().getBean(ReservationRepository.class);
             }
         }
         return reservationRepository;
@@ -82,7 +98,7 @@ public class ReservationServiceImpl implements ReservationService {
         if (roomRepository == null) {
             synchronized (this) {
                 if (roomRepository == null)
-                    roomRepository = ApplicationContext.getBean(RoomRepository.class);
+                    roomRepository = ApplicationContext.getInstance().getBean(RoomRepository.class);
             }
         }
         return roomRepository;
